Move TZDateTime to avoid circular import
